feat: 명확하게 기능 분리 및 네이밍 수정
